+++ resdk/resources/relation.py
@@ -26,8 +26,6 @@ class Relation(BaseResolweResource):
         'collection', 'category', 'partitions', 'unit',
     )
 
-    ALL_PERMISSIONS = ['view', 'edit', 'share', 'owner']
-
     def __init__(self, resolwe, **model_data):
         """Initialize attributes."""
         self.logger = logging.getLogger(__name__)