+++ docs/CHANGELOG.rst
@@ -5,9 +5,9 @@ Change Log
 All notable changes to this project are documented in this file.
 
 
-==========
-Unreleased
-==========
+===================
+12.0.0 - 2019-11-19
+===================
 
 Changed
 -------